pbdrv/bluetooth: Further EV3 RX interrupt safety.
In the case where the RX ring buffer is full, disable interrupts
to prevent an interrupt storm. When consuming bytes from the
RX ring buffer, check if the interrupts are disabled, and if so,
reenable them when the RX ring buffer has space.

Tested by manually setting the ring buffer size to an artificially
short length and observing the interrupts being repeatedly disabled
and enabled. When the ring buffer is an appropriate length, none
of the traffic that the CC2560 module generates during the startup
phase is sufficient to trigger this condition.